Clean up the extra structure (no more union)
git-svn-id: https://twin-e.svn.sourceforge.net/svnroot/twin-e/trunk@114 b6842851-fe11-0410-aa3d-efb4359767e1

+++ twin-e/src/moveActor.c
@@ -1028,6 +1028,6 @@ void ManualRealAngle(int angleFrom, int angleTo, int angleSpeed, timeStruct * an
   numOfStepLong *= angleSpeed;
   numOfStepLong >>= 8;
 
-  angleStruct->numOfStep = numOfStepLong;
+  angleStruct->numOfStep = (short int)numOfStepLong;
   angleStruct->timeOfChange = lba_time;
 }
